Move Welcome Component.
