+++ client/src/app/home/home.component.html
@@ -1,5 +1,1 @@
-<app-welcome></app-welcome>
-<div class="row">
-  <app-submission-list style="width: 100%"></app-submission-list>
-</div>
-<br/>
+<app-submission-list style="width: 100%"></app-submission-list>
