--- client/src/app/home/welcome/welcome.component.html
@@ -1,22 +1,0 @@
-<div>
-  <div class="row">
-    <div class="col-11">
-      <h1 *ngIf="userInfo">Hi {{userInfo?.given_name}},</h1>
-    </div>
-    <div class="col-1"></div>
-  </div>
-  <div class="row">
-    <div class="col-8"
-         *ngIf="(summary$ | async) || {draftSubmissions:0, pendingSubmissions:0, completedSubmissions:0}; let summary">
-      <p>This is your dashboard</p>
-      <p>You currently have {{ summary.draftSubmissions == 0 ? "no" : summary.draftSubmissions}} active draft submissions and {{summary.pendingSubmissions == 0 ? "no" : summary.pendingSubmissions}} pending
-        submissions. You have successfully completed {{summary.completedSubmissions == 0 ? "no" : summary.completedSubmissions}} submissions to the Human Cell
-        Atlas.  {{summary.completedSubmissions == 0 ?  "We're looking forward to your contribution to the future of science." : "Thanks for contributing to the future of science."}}
-      </p>
-    </div>
-    <div class="col-4">
-      <app-new-submission></app-new-submission>
-    </div>
-  </div>
-</div>
-
