+++ client/src/app/submitter/my-projects/my-projects.component.html
@@ -1,6 +1,4 @@
-<div>
-  <button type="button" class="btn btn-success" align="right" (click)="newProject()">Create Project</button>
-</div>
+<app-welcome></app-welcome>
 <br />
 <app-project-list [projects]=projects></app-project-list>
 <mat-paginator [length]="currentPageInfo['totalElements']"
