+++ client/src/app/submitter/welcome/welcome.component.spec.ts
@@ -1,7 +1,7 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {WelcomeComponent} from './welcome.component';
-import {NewSubmissionComponent} from '../new-submission/new-submission.component';
+import {NewSubmissionComponent} from '../../home/new-submission/new-submission.component';
 import {IngestService} from '../../shared/services/ingest.service';
 
 import {RouterTestingModule} from '@angular/router/testing';
