+++ client/src/app/submitter/welcome/welcome.component.ts
@@ -4,6 +4,7 @@ import {Observable} from 'rxjs';
 import {Summary} from './summary';
 import {AaiService} from '../../aai/aai.service';
 import {Profile} from 'oidc-client';
+import {Router} from '@angular/router';
 
 @Component({
   selector: 'app-welcome',
@@ -15,7 +16,7 @@ export class WelcomeComponent implements OnInit {
   userInfo: Profile;
   summary$: Observable<Summary>;
 
-  constructor(public aai: AaiService, private ingestService: IngestService) {
+  constructor(public aai: AaiService, private ingestService: IngestService, private router: Router) {
   }
 
   ngOnInit() {
@@ -23,5 +24,9 @@ export class WelcomeComponent implements OnInit {
       this.userInfo = profile;
     });
     this.summary$ = this.ingestService.getUserSummary();
+  }
+
+  createProject() {
+    this.router.navigate(['projects/new']);
   }
 }
